+++ src/app/shared/types/projection.types.ts
@@ -1,0 +1,46 @@
+/**
+ * Shared projection types used across the application.
+ * This is the canonical source for projection-related type definitions.
+ */
+
+/**
+ * Available projection methods.
+ * Used by ProjectionService, projection.worker, and UI components.
+ */
+export type ProjectionMethod = 'pca' | 'fastmap' | 'isomap' | 'mds' | 'lle' | 'ltsa' | 'tsne' | 'umap' | 'trimap' | 'topomap' | 'sammon';
+
+/**
+ * Result of a projection computation.
+ */
+export interface ProjectionResult {
+  method: ProjectionMethod;
+  positions: Array<{ id: string | number; x: number; y: number }>;
+  computeTime: number;  // milliseconds
+}
+
+/**
+ * Configuration options for projection computations.
+ */
+export interface ProjectionComputeConfig {
+  // t-SNE parameters
+  perplexity?: number;
+  iterations?: number;
+  // UMAP parameters
+  neighbors?: number;
+  minDist?: number;
+  // IsoMap, LLE, LTSA parameters
+  isomapNeighbors?: number;
+  lleNeighbors?: number;
+  ltsaNeighbors?: number;
+  // TriMap parameters
+  trimapWeightAdj?: number;
+}
+
+/**
+ * Progress update during projection computation.
+ */
+export interface ProjectionProgress {
+  method: ProjectionMethod;
+  progress: number;  // 0-100
+  message?: string;
+}
